Crate trillium_caching_headers

Trillium handlers for etag and last-modified-since headers.

This crate provides three handlers: Etag, Modified, and CachingHeaders, as well as a CachingHeadersExt that extends trillium::Headers with some accessors.

Unless you are sure that you don’t want either etag or last-modified behavior, please use the combined CachingHeaders handler.

Structs

CacheControlHeader

A representation of the Cache-Control header.

CachingHeaders

A combined handler that provides both Etag and Modified behavior.

EntityTag

An entity tag, defined in RFC7232

Etag

Etag and If-None-Match header handler

Modified

A handler for the Last-Modified and If-Modified-Since header interaction.

Enums

CacheControlDirective

An enum representation of the Cache-Control directives.

Traits

CachingHeadersExt

Provides an extension trait for both trillium::Headers and also trillium::Conn for setting and getting various parsed caching headers.

Functions

caching_headers

alias for CachingHeaders::new