Trait trillium_caching_headers::CachingHeadersExt

pub trait CachingHeadersExt: Sized {

    // Required methods
    fn etag(&self) -> Option<EntityTag>;
    fn set_etag(&mut self, entity_tag: &EntityTag);
    fn last_modified(&self) -> Option<SystemTime>;
    fn set_last_modified(&mut self, system_time: SystemTime);
    fn cache_control(&self) -> Option<CacheControlHeader>;
    fn set_cache_control(
        &mut self,
        cache_control: impl Into<CacheControlHeader>
    );
    fn if_modified_since(&self) -> Option<SystemTime>;
    fn if_none_match(&self) -> Option<EntityTag>;
    fn set_vary<I, N>(&mut self, vary: I)
       where I: IntoIterator<Item = N>,
             N: Into<HeaderName<'static>>;

    // Provided methods
    fn with_cache_control(
        self,
        cache_control: impl Into<CacheControlHeader>
    ) -> Self { ... }
    fn with_last_modified(self, system_time: SystemTime) -> Self { ... }
    fn with_etag(self, entity_tag: &EntityTag) -> Self { ... }
    fn with_vary<I, N>(self, vary: I) -> Self
       where I: IntoIterator<Item = N>,
             N: Into<HeaderName<'static>> { ... }

}

Provides an extension trait for both [trillium::Headers] and also [trillium::Conn] for setting and getting various parsed caching headers.

Required Methods

fn etag(&self) -> Option<EntityTag>

returns an EntityTag if these headers contain an Etag header.

fn set_etag(&mut self, entity_tag: &EntityTag)

sets an etag header from an EntityTag.

fn last_modified(&self) -> Option<SystemTime>

returns a parsed timestamp if these headers contain a Last-Modified header.

fn set_last_modified(&mut self, system_time: SystemTime)

sets a formatted Last-Modified header from a timestamp.

fn cache_control(&self) -> Option<CacheControlHeader>

returns a parsed CacheControlHeader if these headers include a Cache-Control header. Note that if this is called on a [Conn], it returns the request [Cache-Control] header.

fn set_cache_control(&mut self, cache_control: impl Into<CacheControlHeader>)

sets a Cache-Control header on these headers. Note that this is valid in both request and response contexts, and specific directives have different meanings.

fn if_modified_since(&self) -> Option<SystemTime>

returns a parsed If-Modified-Since header if one exists

fn if_none_match(&self) -> Option<EntityTag>

returns a parsed EntityTag header if there is an If-None-Match header.

fn set_vary<I, N>(&mut self, vary: I)
where I: IntoIterator<Item = N>, N: Into<HeaderName<'static>>,

sets the Vary header to a collection of Into

Provided Methods

fn with_cache_control( self, cache_control: impl Into<CacheControlHeader> ) -> Self

chainable method to set cache control and return self. primarily useful on Conn

fn with_last_modified(self, system_time: SystemTime) -> Self

chainable method to set last modified and return self. primarily useful on Conn

fn with_etag(self, entity_tag: &EntityTag) -> Self

chainable method to set etag and return self. primarily useful on Conn

fn with_vary<I, N>(self, vary: I) -> Self
where I: IntoIterator<Item = N>, N: Into<HeaderName<'static>>,

chainable method to set vary and return self. primarily useful on Conn

Object Safety

This trait is not object safe.